Rename file names to kebab case
