+++ client/src/components/echo-logo.tsx
