+++ client/src/features/landing/layout/layout.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import Page from "../../../components/Page"
+import Page from "../../../components/page"
 
 interface Props {
     children: React.ReactNode
